refactor: migrate Conta relationship from 1:1 to 1:N with account ownership validation

+++ src/services/deletaDados.js
@@ -1,43 +1,51 @@
 import { buscaContaPorUsuarioId } from "./consulta.js";
 
-export const deletaDados = (db, id) => {
+export const deletaUsuario = (db, idUsuario) => {
+  const contas = buscaContaPorUsuarioId(db, idUsuario)
+  const verificaSaldo = contas.some(conta => conta.Saldo > 0)
 
-    const contaAlvo = buscaContaPorUsuarioId(db, id)
+  if (verificaSaldo) {
+    const erro = new Error("A conta não pode ser excluida pois possui saldo")
+    erro.status = 400
+    throw erro
+  }
 
-    const deletaTransacoesOrigem = db.prepare(`
-        DELETE FROM Transacao 
-        WHERE ContaOrigemId = ?
-        `);
-    
-    const deletaTransacoesDestino = db.prepare(`
-        DELETE FROM Transacao 
-        WHERE ContaDestinoId = ?
-        `);
+  const deletaTransacoesOrigem = db.prepare(`
+    DELETE FROM Transacao 
+    WHERE ContaOrigemId = ?
+    `)
+  
+  const deletaTransacoesDestino = db.prepare(`
+    DELETE FROM Transacao 
+    WHERE ContaDestinoId = ?
+    `)
+  
+  const deletaConta = db.prepare(`
+    DELETE FROM Conta 
+    WHERE ContaId = ?
+    `)
+  
+  const queryDeletaUsuario = db.prepare(`
+    DELETE FROM Usuario 
+    WHERE UsuarioId = ?
+    `)
 
-    const deletaConta = db.prepare(`
-        DELETE FROM Conta
-        WHERE ContaId = ?
-        `)
-
-    const deletaUsuario = db.prepare(`
-        DELETE FROM Usuario
-        WHERE UsuarioId = ?
-        `)
-
-    const executaExclusão = db.transaction((uId, cId) => {
-        deletaTransacoesOrigem.run(cId)
-        deletaTransacoesDestino.run(cId)
-        deletaConta.run(cId)
-        deletaUsuario.run(uId)
-    })
-
-    try {
-     executaExclusão(contaAlvo.UsuarioId, contaAlvo.ContaId)
-
-     return contaAlvo.ContaId
-    } catch (error) {
-      const erro = new Error("Não foi possivel realizar a exclusão dos dados: " + error.message)
-      erro.status = 500
-      throw erro
+  const executaExclusao = db.transaction((listaContas, uId) => {
+    for (const conta of listaContas) {
+      deletaTransacoesOrigem.run(conta.ContaId)
+      deletaTransacoesDestino.run(conta.ContaId)
+      deletaConta.run(conta.ContaId)
     }
+    
+    queryDeletaUsuario.run(uId)
+  })
+
+  try {
+    executaExclusao(contas, idUsuario)
+    return idUsuario
+  } catch (error) {
+    const erro = new Error("Não foi possivel realizar a exclusão dos dados")
+    erro.status = 500
+    throw erro
+  }
 }
